move sortByOpen to utils
