+++ packages/web/src/components/CategoryResults.tsx
@@ -3,9 +3,9 @@ import { TResourceCategory, TResourceSubcategory } from "../webTypes";
 import PageBanner from "./PageBanner";
 import React from "react";
 import ResourceList from "./ResourceList";
-import { ResourceSchedule } from "@upswyng/common";
 import SubCategories from "./SubCategories";
 import Typography from "@material-ui/core/Typography";
+import { sortByOpen } from "../utils/schedule";
 import { useParams } from "react-router-dom";
 import useResourcesByCategory from "./useResourcesByCategory";
 import useResourcesBySubcategory from "./useResourcesBySubcategory";
@@ -45,7 +45,7 @@ const CategoryResults = ({
   const status = subcategoryStub
     ? subcategoryResourcesStatus
     : categoryResourcesStatus;
-  const resources = ResourceSchedule.sortByOpen(
+  const resources = sortByOpen(
     subcategoryStub ? subcategoryResources : categoryResources
   );
